requirements.txt
# requirements. 
# Use
# pip3 install -r requirements.txt

Flask
python-docx
docx2pdf
openpyxl
pythoncom
